src/requirements.txt
matplotlib==3.1.0
cycler==0.10.0
kiwisolver==1.1.0
pyparsing==2.4.0
python-dateutil==2.8.0
numpy==1.16.4
six==1.12.0
setuptools==41.0.1
umap-learn==0.4.0rc1
scipy==1.3.0
scikit-learn==0.21.2
tbb==2020.0.133
numba==0.44.1
joblib==0.13.2
llvmlite==0.29.0